Refactor plex_connect for maintainability and contributions
Addresses style fixes and better segregation of functionalities to make
the plugin easier to maintain and for others to contribute to.

The monolithic player_remote.py is split into focused modules:

- server.py         — core HTTP server, routing, lifecycle
- timeline.py       — timeline XML building and broadcasting
- playback.py       — playback control command handlers
- queue_commands.py — play queue HTTP command handlers
- queue_sync.py     — background queue loading and MA↔Plex sync logic

Also promotes the plugin stage from alpha to beta, and fixes several
bugs discovered during the refactor: a missing _updating_from_plex guard
in handle_create_play_queue, a race condition in background queue loading,
and shuffle changes not being synced back to the Plex play queue.
